refactor: split high-conflict files into focused modules
- Split helpers.ts (316 lines, 13 exports) into 6 domain modules:
  recipeScaling, shoppingList, equipment, versioning, eventUtils, ratings
  helpers.ts retained as barrel re-export for backward compatibility
- Extract useRecipeFiltering hook from RecipeLibrary.tsx (filter/sort state)
- Extract useFeedbackForm hook from EventFeedback.tsx (form state + handlers)
- Extract useAppData hook from App.tsx (12 query/mutation hooks + handlers)

All existing imports continue to work via barrel re-exports.
Build passes, 68 tests pass.
